+++ src/config.ts
@@ -3,6 +3,7 @@ import * as vscode from "vscode";
 const DEFAULT_DELAY_MS = 30000;
 const ENABLED_KEY = "enabled";
 
+export type AutoPull = AutoPush | "onPush";
 export type AutoPush = "onCommit" | "afterDelay" | "off";
 export type CommitValidationLevel = "error" | "warning" | "none";
 
@@ -14,8 +15,14 @@ export default {
   get autoCommitDelay(): number {
     return config().get("autoCommitDelay", DEFAULT_DELAY_MS);
   },
+  get autoPull(): AutoPull {
+    return config().get("autoPull", "onPush");
+  },
+  get autoPullDelay(): number {
+    return config().get("autoPullDelay", DEFAULT_DELAY_MS);
+  },
   get autoPush(): AutoPush {
-    return config().get("autoPush", "off");
+    return config().get("autoPush", "onCommit");
   },
   get autoPushDelay(): number {
     return config().get("autoPushDelay", DEFAULT_DELAY_MS);
@@ -26,6 +33,9 @@ export default {
   get commitValidationLevel(): CommitValidationLevel {
     return config().get("commitValidationLevel", "error");
   },
+  get commitOnClose() {
+    return config().get("commitOnClose", true);
+  },
   get enabled() {
     return config().get(ENABLED_KEY, false);
   },
@@ -34,5 +44,8 @@ export default {
   },
   get filePattern() {
     return config().get("filePattern", "**/*");
+  },
+  get pullOnOpen() {
+    return config().get("pullOnOpen", true);
   },
 };