+++ src/store.ts
@@ -2,5 +2,6 @@ import { observable } from "mobx";
 
 export const store = observable({
   enabled: false,
+  isPulling: false,
   isPushing: false,
 });